pyproject.toml
[tool.poetry]
name = "chat-widget"
version = "0.1.0"
description = ""
authors = ["Raghu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.4"
authlib = "^1.3.2"
itsdangerous = "^2.2.0"
httpx = "^0.27.2"
python-multipart = "^0.0.12"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
